Add CMakeLists.txt and git-pre-commit-hook, rework git-clang-format-cmake.py
Adding CMakeLists.txt with add_subdirectory enables two features, the
first is installing a git pre-commit hook (a thin wrapper around
git-clang-format-cmake.py with correct invocation) to block any commits
where source files have not been formatted, and a custom cmake target
called 'format' which is responsible for formatting all edited files
(again calling git-clang-format-cmake.py with the alternative invocation).

At cmake time the paths to git, python, and clang-format are searched,
failure to find these executables on your path will results in a cmake
error. Alternatively it is possible to set one or all of GCFC_GIT_PATH,
GCFC_PYTHON_PATH, and GCFC_CLANGFORMAT_PATH manually, a warning will be
produced if the file path does not exist. The decision not to cause a
cmake error is to enable the path to any or all of these dependencies
to be built in your cmake tree, however this is currently untested and
is considered a future feature at this time.

Minor changes were required to the git-clang-format-cmake.py script to
enable correct behaviour when invoked from the pre-commit hook and the
cmake format target.