Kill descendants on exit
This is a best effort attempt to kill all descendants on exit. We will
reads the /proc file system to find out all of our descendants.

There are race conditions that not possible for us to resolve. For example,
new processes can be created while we are reading /proc. So there might
be left overs.