object: fix the problem in 6ecf02709511
Firstly, make sure di_type_conversion returns the same pointer as the
input pointer if no conversion happened. This is to make it easier to
identify the cases where the output value needs to be freed. (If no
conversion happened, the caller is responsible of freeing the value).

Then, make sure values returned by di_type_conversion is properly freed.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>