object: add the di_variant type
Current there is no support at a type system level for a singular
variant value, you can only have a tuple of variant values. This is
inconvenient for functions which want to have a variant return type.

So we added a variant type.

But di_variant and di_tuple aren't completely orthogonal, so next step
we will replace di_tuple with di_array of di_variant.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>