plugins: lua: remove unnecessary acces of the script object
The script object was used to keep track of di_objects in the lua
scripts, so the objects can be freed when script is unloaded.

How object are freed have since changed, but the access of script object
was not removed, even though the script object was just retrived then
popped.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>