object: change how listeners reference things
Now the listeners no longer have strong references to the event
sources and the event sources no longer have one to the listeners.
They all use weak refs now.

Which means you need to keep a strong ref to both the listen handle and
the event source to keep them alive. Or else you won't be getting
anything.

di_listen_to_once is removed. As you would have to do the same thing
as a repeating listeners to keep them alive, and dispose of them anyway.

The plugins have been updated to do this. The lua plugin has to do
something special to avoid ref cycles and to make the user's life
easier. See the comment at the top of plugins/lua/lua.c for more
information.

Also I tried to keep object reference as members of the owning object,
to take advantage of the automatic memory management.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>