[general]: strings are now conceptually immutable
To avoid the confusion when a callee tries to modify the string passed
to it and expect it to be reflected to the caller. The caller could have
passed a string literal and this won't work, as string literal will be
cloned previously.

Now, all strings are immutable. So there is no such confusion. And this
also saves us a copy when passing borrowed string literal as borrowed
string.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>