object: remove di_destroy_object from public api
Listeners cannot be unilaterally removed anymore, they can only be
detached when strong ref drop to 0.

However, there are sometimes needs to free all the resources held by an
object. Just calling the dtor is not enough, and there can still be
members holding resources. And we can't ask the dtor to remove members,
that would cause lots of code duplication.

So add a new di_finalize_object, which calls the dtor and remove all
members.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>