object: let char * mean DI_TYPE_STRING
This way there is a way for functions to state that they are returning a
string literal, so they can avoid unnecessary copies. Currently
functions returning "const char *" are always detected as returning
owned strings, so they have to copy what they return.

This has some complications, in that DI_TYPE_STRING is a const char * in
arguments, but is a char * in returned values.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>