plugins: dbus: fix cycles in DBusWatch
Previously dbus watch functions hold a strong reference to the
connection object, to use in dbus_add_watch/remove_watch callbacks. This
creates a cycle, because the watch functions and the connection object
will keep each other alive.

So we hold a weak reference instead. This is fine as the watch callbacks
are only called when the connection is alive.

Only exception is during the 2 step connection shutdown. Because we
used to only close the connection after the connection object is
destructed, and closing the connection calls the remove_watch callback.
So now we remove the watch functions in the first half of connection
shutdown, to prevent remove_watch from being called, as there is no need
anyway.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>