plugins: dbus: fix memory leak in case di.exit is called
The dbus connection object isn't freed until we return to the mainloop,
but if someone calls di.exit/quit before that, we never received the
signal telling us we returned to the mainloop.

So instead of relying on a signal is delievered, we instead put the
clean up code in a dtor of a temporary object then add it to anonymous
roots. If the prepare signal is delievered, we remove it from the roots
and its dtor is called; if not deai will clear all roots and call its
dtor, so its dtor is called either way.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>