plugins: lua: fix object tracking
Weak lua table entries might die before __gc for that entry is called.
In that case a new object_to_ref entry is created and a new reference is
created. Later in __gc for the old reference, that entry will be removed
even though the new reference is still alive.

This commit reworks the object tracking mechanism to be simpler and
fixes this bug.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>