builtin: event: no longer guarantee the delievery of "prepare"
Previously we keep the core object alive so in case it dies before
prepare is delievered. But that's not enough to guarantee the delievery
of "prepare", because someone could call di_quit/di_exit, and "prepare"
won't be delievered in that case.

Our code has already been fixed to not rely on "prepare" to always be
delivered, so remove this guarantee.

This fixes a reference cycle.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>