object: add an empty object type
This is to avoid an allocation when converting nil -> object. This is
initially for being able to use empty Lua tables as either an object or
an array.

Now the empty object fills that role. This have added benefit that the
object actually is the empty lua table. Previously a completely
unrelated object will be created.

Also di_type_conversion now handles nil too.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>