object: fix use-after-free in garbage collection
Previously we first increment reference counts of objects reachable
from one unreferred roots, then finalize all of those objects.

The thought is that, because there are reference cycles, a object's
reference count could drop to 0 while we are finalizing it (because we
call finalize when the object's reference count hadn't reached 0 yet),
causing finalize to be called recursively, which is bad.

However this is not enough: because objects can have destructors. And a
destructor can cause another object that will be freed by the garbage
collector later, via a later unreferred root, to be freed early. This
can be possible by upgrading a weak reference to the other object. (If
the object holds a strong reference instead of a weak one, then the
other object will be reachable from the current unreferred root, and
will have had its reference count incremented).

More specifically, suppose this is the reference graph

	unreferred
	roots list
	|
	v
	object_A --(weak reference)-> object_B
	|
	v
	object_B -> object_C -> object_B (reference cycle)

While finalizing objects reachable from object_A, we will not increment
the reference count of object_B, because of the weak ref. But the
destructor of object_A can upgrade the weak ref to a strong ref of
object_B, then get object_C, then remove object_B from object_C, then
drop its own strong ref and weak ref, causing object_B to be freed. Then
the garbage collector will use-after-free when it advance its roots
iterator to object_B.

So basically we have to increment reference counts of all the objects
that will be freed first, before trying to free any of them.

Also make sure mark == 0 outside of collect_garbage and dump_objects.
Useful for later.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>