object: fix garbage collection again
d0250baa62e tried to address use-after-free in garbage collector,
however it's both insufficient and excessive. It refs way more objects
than necessary. Yet it only refs objects that are marked for
destruction, whereas objects that are not marked can still be freed by
user dtor, causing use-after-free.

The real problem is actually simpler: we have references that are not
accounted for in the reference count - the unreferred_object list. So
we really just need to ref objects in the list.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>