main: load plugins from environment variable
The command line can only be used to do one thing. If we need to load
the lua plugin from somewhere non-standard, right now we have to do it
using the command line, which stops us from doing anything useful. So
instead deai now automatically loads plugins specified in environment
variables.

This solves the problem when we need to run deai from the build folder,
e.g. for unit testing.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>