callable: split closure into 2 halves
Currently the closure trampoline does 2 things: 1. prepend the captured
variable onto the list of arguments; 2. do type conversions and call the
actual function with libffi.

This change split this 2 steps. A raw closure type is created that only
does the first step, which is intended to be used by functions that
don't need to be called via libffi.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>