callable: move exception handling output di_closure
Previously every di closure implicitly catches exceptions thrown by the
target function called. Essentially there is no chance for user code to
catch and handle any exceptions.

This commit got rid of that exception catch, and only catches exceptions
at the "bottom of the stack", so it's either 1) when calling a signal
listener, or 2) in main() when calling the initial method.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>