plugins/lua: passthrough lua errors and exceptions
When a lua function called from C generates an error, it is passed
through as a C++ exception; and when a C++ exception is thrown from a
C function called by lua, a lua error is generated. This way we make
sure errors are propagated all the way through until someone catches it.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>