plugins/lua: can't assume stack top is 0 when coroutine exists
lua coroutines exist on a different lua state from the main lua state,
when the main lua state is resuming a coroutine, its stack might not be
empty. so if:

lua_load_script -> lua script -> coroutine.resume ->
lua script calls a function in di.lua that uses the main state.

that di.lua function will see a non-empty lua state.

so instead of making sure we always leave the lua stack empty when
returning from these functions, we make sure we always leave the stack
how we found it.