spawn: let's not mess with ptys
If we open a master pty for children, then they will get SIGHUPs when
the master pty is closed. If we have grandchildren, they will all die
because of this. This breaks spawn.run(rofi) then starting programs from
rofi.

Let's just detach children from current session and call it a day.